Prevent double promise resolution
It is possible to call requestCredentials multiple times in quick
succession, and therefore the first Promise will be called twice. It is
also possible that the API has some unintended workings. Fixed by
ensuring that the promise only gets called once. The first request
fails and performs an error log in logcat, look out for that!